requirements.txt
Django==2.0
django-cors-headers==3.4.0
djangorestframework==3.11.0
gunicorn==19.7.1
Pillow==7.2.0
whitenoise==3.3.1

